+++ front/src/components/PixelIdentifierType.vue
