+++ front/src/views/FindEmployeeView.vue
@@ -1,10 +1,10 @@
 <template>
     <FeedBack></FeedBack>
-    <PixelIDChoice></PixelIDChoice>
+    <Identifier></Identifier>
     <PixelUserInput></PixelUserInput>
 </template>
 <script setup lang="ts">
 import PixelUserInput from '@/components/PixelUserInput.vue'
-import PixelIDChoice from '@/components/PixelIDChoice.vue'
+import Identifier from '@/components/PixelIdentifierType.vue'
 import FeedBack from '@/components/PixelIdentifierFeedback.vue';
 </script>
